Remove daemon package
